Make appending events a procedure
The main driving factor being the ability to commit the transaction
after every loop iteration when adding multilpe events.

Previously, when two batches of events were being added at the same
time, the events would get also batched up and emitted once the final
transaction (wrapping the whole function) is commited.

If two calls to the function were adding events [a,b] and [c,d], the
sequence numbers could be (correctly)
(a, 1), (c, 2), (b, 3), (d, 4)

But then notifications will be (assuming the [a,b] batch finishes and
commits first), in order

(a, 1)
(b, 3)
(c, 2)
(d, 4)

With the idempotent projections this means that (c,2) gets ignored,
which is obviously bad.

Comitting after every loop would emit the notifications immediatelly
after the event was stored, and should hopefully result in in-order
notifications (no longer batched up) as well.