feat(examples): bare minimum "light switch" example (#322)
This example application is a bare minimum example of the core
functionality. Persistence is not enabled, nor is logging, nor any kind
of user interface. It just starts, runs some commands, queries the final
state and prints it.

It models a "light switch". You can "install" the switch, then turn it
"off" or "on" using commands. You can issue a query to get the current
state of a switch.